refactor: Move fixtures to "src/" to encourage a single point to gather them
