+++ src/profile/Profile.mdx
@@ -7,7 +7,7 @@ import { ProfileItems } from "./ProfileItems";
 import { AlertsDropdown } from "./AlertsDropdown";
 import { InboxDropdown } from "./InboxDropdown";
 import { UserAvatarDropdown } from "./UserAvatarDropdown";
-import { notifications } from "../notifications/__tests__/__fixtures__/notifications";
+import { notifications } from "../__tests__/__fixtures__/notifications";
 
 # Profile
